--- android/app/src/main/cpp/CMakeLists.txt
@@ -1,30 +1,0 @@
-#
-# Copyright 2023 The Nakama Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-file(
-    GLOB_RECURSE srcs
-    LIST_DIRECTORIES false
-    "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/*.h"
-)
-
-add_library(android-ca OBJECT ${srcs})
-
-target_include_directories(android-ca PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
-
-target_link_libraries(android-ca
-                      android
-                      nakama-sdk-core-misc)
-
