+++ impl/httpCurl/CMakeLists.txt
@@ -5,6 +5,8 @@ file(
         "${CMAKE_CURRENT_SOURCE_DIR}/*.h"
 )
 
+add_subdirectory(android)
+
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
 add_library(nakama-impl-http-libcurl OBJECT ${srcs})
