+++ impl/httpCurl/android/AndroidCA.h
