+++ impl/httpCurl/android/CMakeLists.txt
@@ -1,0 +1,13 @@
+file(
+        GLOB_RECURSE srcs
+        LIST_DIRECTORIES false
+        "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp"
+        "${CMAKE_CURRENT_SOURCE_DIR}/*.h"
+)
+
+add_library(android-ca OBJECT ${srcs})
+target_include_directories(android-ca PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_link_libraries(android-ca
+    PRIVATE android nakama-sdk-core-misc
+)
